tests/submission/Infraware/Forms/contents/Forms/textarea_value.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>textarea_value</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action=""
		  name="input_form">
	<p><textarea id='textarea_id'></textarea></p>
	</form>

	<script>
		
		var textarea = document.getElementById("textarea_id");
		textarea.value = 'val';

		if (typeof(textarea.value) == "string") {
			test( function() {
				assert_equals(textarea.value, "val", "value attribute is not correct.");
			});
		} else {
			test( function() {
				assert_unreached("value attribute is not exist.");
			});
		}

	</script>

 </body>
</html>
